Refactor environment imports across services and configuration files to use the unified environment module, enhancing maintainability and consistency in API URL management.
