+++ src/app/core/services/admin.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { API_URLS } from '../constants/apiurl.constants';
 import { HttpClient } from '@angular/common/http';
 import { IAdmin } from '../models/admin.model';
