+++ src/app/core/services/ai-chat.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { IResponse } from '../models/httpResponse.models';
 import { ITrip } from '../models/trip.model';
