+++ src/app/core/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, catchError, map, of, tap } from 'rxjs';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { IOtpVerify, ISignup, Ilogin } from '../models/auth.models';
 import { CookieService } from 'ngx-cookie-service';
 import { API_URLS } from '../constants/apiurl.constants';
