+++ src/app/core/services/conversation.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { API_URLS } from '../constants/apiurl.constants';
 import IConversation from '../models/conversation.model';
 import { Observable } from 'rxjs';
