+++ src/app/core/services/map.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { environment } from '../../../environments/environment.development';
+import { environment } from '../../../environments/environment';
 import { IPlaceSuggestion } from '../models/mapService.models';
 import { API_URLS } from '../constants/apiurl.constants';
